+++ mst prim.cpp
@@ -1,10 +1,10 @@
 //Usage prim(source). mst holds new tree
 // adjlst stores <weight,node>
-vector<pair<long long int,int> > adjlst[MAX],mst[MAX];
+vector<pair<int,int> > adjlst[MAX],mst[MAX];
 int par[MAX],vis[MAX],n,m;
-long long int dst[MAX];
+int dst[MAX];
 vector<int> path;
-long long int prim(int u)
+int prim(int u)
 {
 	int i;
 	memset(par,0,sizeof(par));
@@ -14,8 +14,8 @@ long long int prim(int u)
 		mst[i].clear();
 	}
 	memset(vis,0,sizeof(vis));
-	long long int sm=0;
-	priority_queue<pair<long long int,int>, vector<pair<long long int,int> >, greater<pair<long long int,int> > > pq;
+	int sm=0;
+	priority_queue<pair<int,int>, vector<pair<int,int> >, greater<pair<int,int> > > pq;
 	pq.push(make_pair(0LL,u));
 	par[u]=u;
 	dst[u]=0;
